Added Additional Cams, Limits, Enhancements.
Introduced SideCam and BirdsEyeCam to provide additional views for the
user.
Added limits to position and rotation of the endoscope, e.g. the user
can not get out of the body in a specific scenario without making any
additional moves.
Added GUI's for all the different screens as well as statistics uppon
scenario completion and in-scenario stats, i.e. time, length, etc.